# backend/openapi/credence_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Credence AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/reputation/{wallet}": {
      "get": {
        "tags": [
          "Reputation"
        ],
        "summary": "Get Reputation",
        "description": "Retrieves the authoritative reputation profile, trust score, rating,\nand history of versioned profiles for a borrower address.",
        "operationId": "get_reputation_reputation__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/policies": {
      "get": {
        "tags": [
          "Policy Engine"
        ],
        "summary": "List Policies",
        "description": "Retrieves the list of all registered credit policies.",
        "operationId": "list_policies_policies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Policy"
                  },
                  "type": "array",
                  "title": "Response List Policies Policies Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Policy Engine"
        ],
        "summary": "Create Policy",
        "description": "Registers a new credit evaluation policy in the registry.",
        "operationId": "create_policy_policies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Policy"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/policies/evaluate": {
      "post": {
        "tags": [
          "Policy Engine"
        ],
        "summary": "Evaluate Policy",
        "description": "Evaluates a specific credit policy against a wallet parameters context.",
        "operationId": "evaluate_policy_policies_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluatePolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyEvaluation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/policies/{policy_id}": {
      "delete": {
        "tags": [
          "Policy Engine"
        ],
        "summary": "Delete Policy",
        "description": "Removes a policy record from the database registry.",
        "operationId": "delete_policy_policies__policy_id__delete",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/policies/{policy_id}/export": {
      "get": {
        "tags": [
          "Policy Engine"
        ],
        "summary": "Export Policy",
        "description": "Exports a portable credit policy template including SHA-256 validation checksum.",
        "operationId": "export_policy_policies__policy_id__export_get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Policy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/graph/{wallet}": {
      "get": {
        "tags": [
          "Reputation Graph Layer"
        ],
        "summary": "Get Reputation Graph",
        "description": "Returns nodes and edges mappings for a wallet.",
        "operationId": "get_reputation_graph_graph__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustGraph"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/{wallet}/insights": {
      "get": {
        "tags": [
          "Reputation Graph Layer"
        ],
        "summary": "Get Reputation Insights",
        "description": "Assembles AI-narrated graph relationships insights.",
        "operationId": "get_reputation_insights_graph__wallet__insights_get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/{wallet}/connections": {
      "get": {
        "tags": [
          "Reputation Graph Layer"
        ],
        "summary": "Get Reputation Connections",
        "description": "Lists direct connections with active protocols.",
        "operationId": "get_reputation_connections_graph__wallet__connections_get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph/{wallet}/trust-flow": {
      "get": {
        "tags": [
          "Reputation Graph Layer"
        ],
        "summary": "Get Trust Flow Propagation",
        "description": "Traces propagation sources and paths mapping trust coefficients.",
        "operationId": "get_trust_flow_propagation_graph__wallet__trust_flow_get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustFlowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/governance/policy": {
      "post": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Register Policy Rule",
        "operationId": "register_policy_rule_governance_policy_post",
        "parameters": [
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePolicyRequest": {
        "properties": {
          "policy_name": {
            "type": "string",
            "title": "Policy Name"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0.0"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/PolicyRule"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "policy_name",
          "protocol",
          "rules"
        ],
        "title": "CreatePolicyRequest"
      },
...
      "EvaluatePolicyRequest": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "policy_id"
        ],
        "title": "EvaluatePolicyRequest"
      },
...
      "GraphEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "relationship": {
            "type": "string",
            "title": "Relationship"
          },
          "strength": {
            "type": "integer",
            "title": "Strength"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "relationship",
          "strength",
          "created_at"
        ],
        "title": "GraphEdge"
      },
      "GraphNode": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "node_type": {
            "type": "string",
            "title": "Node Type"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "trust_score": {
            "type": "integer",
            "title": "Trust Score"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "node_type",
          "label",
          "trust_score",
          "metadata"
        ],
        "title": "GraphNode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Policy": {
        "properties": {
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_name": {
            "type": "string",
            "title": "Policy Name"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/PolicyRule"
            },
            "type": "array",
            "title": "Rules"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "policy_id",
          "policy_name",
          "protocol",
          "version",
          "rules",
          "created_at"
        ],
        "title": "Policy"
      },
      "PolicyEvaluation": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "matched_rules": {
            "type": "integer",
            "title": "Matched Rules"
          },
          "failed_rules": {
            "type": "integer",
            "title": "Failed Rules"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "policy_id",
          "passed",
          "matched_rules",
          "failed_rules",
          "score",
          "reason"
        ],
        "title": "PolicyEvaluation"
      },
      "PolicyRule": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "operator": {
            "type": "string",
            "title": "Operator"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "field",
          "operator",
          "value"
        ],
        "title": "PolicyRule"
      },
...
      "TrustFlowResponse": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "trust_sources": {
            "items": {
              "$ref": "#/components/schemas/TrustSource"
            },
            "type": "array",
            "title": "Trust Sources"
          },
          "trust_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trust Path"
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "trust_sources",
          "trust_path"
        ],
        "title": "TrustFlowResponse"
      },
      "TrustGraph": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "network_score": {
            "type": "integer",
            "title": "Network Score"
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "nodes",
          "edges",
          "network_score"
        ],
        "title": "TrustGraph"
      },
      "TrustSource": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "impact": {
            "type": "string",
            "title": "Impact"
          }
        },
        "type": "object",
        "required": [
          "source",
          "impact"
        ],
        "title": "TrustSource"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}